feat: move K8s models to k8s
