+++ bases/renku_data_services/k8s_cache/main.py
@@ -5,10 +5,10 @@ import logging
 
 import kr8s
 
+from renku_data_services.k8s.models import ClusterId
 from renku_data_services.k8s_cache.config import Config
 from renku_data_services.k8s_watcher import K8sWatcher
 from renku_data_services.k8s_watcher.db import Cluster, k8s_object_handler
-from renku_data_services.k8s_watcher.models import ClusterId
 from renku_data_services.notebooks.constants import AMALTHEA_SESSION_KIND, JUPYTER_SESSION_KIND
